abstracting out my routing into a router.js file that will sit in the app director.
